refactor: move SocialButtons into SharedComponents
