+++ src/components/BioModal/BioModal.tsx
@@ -17,7 +17,7 @@ import Grid, { GridArea } from "@/components/SharedComponents/Grid";
 import MaxWidth from "@/components/SharedComponents/MaxWidth";
 import ModalDialog from "@/components/ModalDialog";
 import useModalDialog from "@/components/ModalDialog/useModalDialog";
-import SocialButtons from "@/components/SocialButtons";
+import SocialButtons from "@/components/SharedComponents/SocialButtons";
 import BoxBorders from "@/components/SpriteSheet/BrushSvgs/BoxBorders";
 import Svg from "@/components/SharedComponents/Svg";
 import { Heading, P } from "@/components/SharedComponents/Typography";
