+++ src/components/Menu/Menu.tsx
@@ -7,12 +7,12 @@ import { FocusOn } from "react-focus-on";
 import { useMenuContext } from "../../context/Menu/";
 import { OakColorName, PixelSpacing } from "../../styles/theme/types";
 import Logo from "../Logo";
-import SocialButtons from "../SocialButtons";
-import { OAK_SOCIALS } from "../SocialButtons/SocialButtons";
+import { OAK_SOCIALS } from "../SharedComponents/SocialButtons/SocialButtons";
 import SideBarSignpost from "../SideBarSignpost/SideBarSignpost";
 
 import MenuBackdrop from "./MenuBackdrop";
 
+import SocialButtons from "@/components/SharedComponents/SocialButtons";
 import Svg from "@/components/SharedComponents/Svg";
 import IconButton from "@/components/SharedComponents/Button/IconButton";
 import flex from "@/styles/utils/flex";
